requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
appdirs==1.4.4
appnope==0.1.2
asttokens==2.0.5
backcall==0.2.0
blas==1.0
bottleneck==1.3.5
brotli==1.0.9
brotli-bin==1.0.9
brotlipy==0.7.0
bzip2==1.0.8
ca-certificates==2023.01.10
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==7.1.2
comm==0.1.2
contourpy==1.0.5
cryptography==39.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
ffmpeg==4.2.2
flit-core==3.6.0
fonttools==4.25.0
freetype==2.12.1
gettext==0.21.0
giflib==5.2.1
gmp==6.2.1
gnutls==3.6.15
icu==68.1
idna==3.4
imbalanced-learn==0.10.1
imblearn==0.0
importlib_resources==5.2.0
ipykernel==6.19.2
ipython==8.10.0
jedi==0.18.1
joblib==1.1.1
jpeg==9
jupyter_client==7.4.9
jupyter_core==5.2.0
kiwisolver==1.4.4
lame==3.100
lcms2==2.12
lerc==3.0
libblas==3.9.0
libbrotlicommon==1.0.9
libbrotlidec==1.0.9
libbrotlienc==1.0.9
libcblas==3.9.0
libcxx==14.0.6
libdeflate==1.17
libffi==3.4.2
libgfortran==5.0.0
libgfortran5==11.3.0
libiconv==1.16
libidn2==2.3.1
liblapack==3.9.0
libopenblas==0.3.21
libopus==1.3
libpng==1.6.39
libsodium==1.0.18
libtasn1==4.16.0
libtiff==4.5.0
libunistring==0.9.10
libvpx==1.10.0
libwebp==1.2.4
libwebp-base==1.2.4
libxml2==2.9.14
llvm-openmp==14.0.6
lz4-c==1.9.4
matplotlib==3.7.1
matplotlib-base==3.7.1
matplotlib-inline==0.1.6
munkres==1.1.4
ncurses==6.4
nest-asyncio==1.5.6
nettle==3.7.3
numexpr==2.8.4
numpy==1.23.5
numpy-base==1.23.5
openh264==1.8.0
openssl==1.1.1
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.4.0
pip==23.0.1
platformdirs==2.5.2
pooch==1.4.0
prompt-toolkit==3.0.36
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.2
pycparser==2.21
pygments==2.11.2
pyopenssl==23.0.0
pyparsing==3.0.9
pysocks==1.7.1
python==3.8.16
python-dateutil==2.8.2
pytorch==1.12.0
pytz==2022.7
pyzmq==23.2.0
readline==8.2
requests==2.28.1
scikit-learn==1.2.2
scipy==1.10.0
setuptools==65.6.3
six==1.16.0
sqlite==3.41.1
stack_data==0.2.0
threadpoolctl==2.2.0
tk==8.6.12
torchaudio==0.12.0
torchvision==0.13.0
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.5
wheel==0.38.4
x264==1!152.20180806
xz==5.2.10
zeromq==4.3.4
zipp==3.11.0
zlib==1.2.13
zstd==1.5.2
